requirements.txt
# --- FastAPI app ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# --- Auth / JWT ---
PyJWT==2.9.0
descope==1.7.9           # <- last available on PyPI as of now

# --- Core numerics / vector search (keep NumPy < 2 for faiss) ---
numpy==1.26.4
faiss-cpu==1.7.4
packaging==24.1

# --- Embeddings / ML ---
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
transformers==4.43.4
sentence-transformers==2.7.0

# --- Files / docs / PDF ---
PyMuPDF==1.24.9          # import fitz
python-pptx==1.0.2        # import pptx
python-docx==1.1.2        # import docx
markdown2==2.5.0
pdfkit==1.0.0             # needs wkhtmltopdf system binary
python-dateutil==2.9.0.post0
ics==0.7.2

# --- OpenAI SDK ---
openai>=1.30,<2
httpx<0.27

